requirements.txt
numpy>=1.20.0
scipy>=1.7.0
scikit-learn>=1.0.0
scikit-image>=0.18.0
matplotlib>=3.3.0
seaborn>=0.11.0
torch>=1.10.0
torchvision>=0.11.0
pyyaml>=5.4.0
tqdm>=4.62.0
Pillow>=8.0.0

